using intptr_t as the size_type for containers refs #92 fixes MSVC "possible loss of data" warning when compiling for 64bits

+++ src/blackmisc/collection.h
@@ -42,7 +42,7 @@ namespace BlackMisc
         typedef typename Iterators::ConstForwardIterator<T> const_iterator;
         typedef const_iterator iterator; // can't modify elements in-place
         typedef ptrdiff_t difference_type;
-        typedef int size_type;
+        typedef intptr_t size_type;
         //! @}
 
         /*!
